# Check for object existence in parallel