@fyrchik 2022-02-28
Vegman results (each metabase is on a separate SSD, each blobstor on a separate HDD). Only 4 disks are there, so I skipped test with 8 shards. `pool` version is the same as this PR but with a separate pool of goroutines (via ants).

```
name \ time/op      old          new           pool
Exists/2_shards-40  52.3µs ±23%   91.6µs ± 6%   84.2µs ± 1%
Exists/4_shards-40  72.7µs ±11%  121.8µs ± 8%  116.9µs ± 5%

name \ alloc/op     old          new           pool
Exists/2_shards-40  5.00kB ± 0%   5.03kB ± 0%   5.16kB ± 0%
Exists/4_shards-40  9.89kB ± 0%   9.93kB ± 1%  10.19kB ± 0%

name \ allocs/op    old          new           pool
Exists/2_shards-40     112 ± 1%      115 ± 1%      117 ± 1%
Exists/4_shards-40     207 ± 1%      211 ± 1%      216 ± 1%
```

Pprof CPU profiles:
[cpuold.out.gz](https://github.com/nspcc-dev/neofs-node/files/8151753/cpuold.out.gz)
[cpunew.out.gz](https://github.com/nspcc-dev/neofs-node/files/8151752/cpunew.out.gz)
[cpupool.out.gz](https://github.com/nspcc-dev/neofs-node/files/8151909/cpupool.out.gz)

It baffled at me at first, that the relative results are _worse_ in this case. I think this is because synchronization costs are comparable to the `exists` check execution time.

So I think the change in it's current form is not worth it.